moving HightPrecisionTime back to values folder since there is no package object anymore
